Make importing more generic
Plugins can now easily register own fact importers.
